app/requirements.txt
dateparser==1.2.0
datefinder==0.7.3
joblib==1.4.2
langdetect==1.0.9
nltk==3.9.1
pdf2image==1.17.0
pdfplumber==0.11.5
pytesseract==0.3.13
python-dotenv==1.0.1
spacy==3.8.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0.tar.gz
https://github.com/explosion/spacy-models/releases/download/nl_core_news_sm-3.8.0/nl_core_news_sm-3.8.0.tar.gz
streamlit==1.41.1
streamlit_pdf_viewer==0.0.20
transformers==4.47.1
unstructured_ingest==0.3.11
scikit-learn==1.6.0
